Tidy Up code
Combined all the import function to come under import_entities.py in functions to be used in telegram_handlers.py and monitor.py

Updated postman calls

Removed irrelevant files
