+++ NGSI-LD-SG-Datamall/mylibs/weather.py
@@ -1,27 +1,97 @@
 import mylibs.constants as constants
-from landtransportsg import Traffic
 import requests
-import urllib.parse
-
 import json
-from requests.exceptions import RequestException, HTTPError
 from ngsildclient import Client, Entity, SmartDataModels
-from datetime import datetime
+from datetime import datetime, timedelta
 
+# ============================ Weather Forecast ======================
 
-API_KEY = constants.DATAMALL_API_KEY
+TWO_HOUR_WEATHER_URL = "https://api-open.data.gov.sg/v2/real-time/api/two-hr-forecast"
+PSI_URL = "https://api-open.data.gov.sg/v2/real-time/api/psi"
+TEMPERATURE_URL = "https://api-open.data.gov.sg/v2/real-time/api/air-temperature"
+RAINFALL_URL = "https://api-open.data.gov.sg/v2/real-time/api/rainfall"
+WIND_URL = "https://api-open.data.gov.sg/v2/real-time/api/wind-direction"
+UVI_URL = "https://api-open.data.gov.sg/v2/real-time/api/uv"
+PM25_URL = "https://api-open.data.gov.sg/v2/real-time/api/pm25"
+
 ctx = constants.ctx
 broker_url = constants.broker_url
 broker_port = constants.broker_port  # default, 80
 temporal_port = constants.temporal_port  # default 1026
 broker_tenant = constants.broker_tenant
 
+# Convert to NGSI-LD
+"""
+Weather Forecast
+- id - from SDM
+- validFrom - From SDM
+- validTo - From SDM
+- weatherType - From SDM 
+- validity - From SDM 
+
+Example NEA data return:
+            "area": "Tampines",
+            "forecast": "Partly Cloudy (Night)"
+
+"""
+
+def get_two_hour_weather():
+    response = requests.get(
+        TWO_HOUR_WEATHER_URL,
+        headers={
+            "Content-Type": "application/json",
+            "User-Agent": "Mozilla/5.0",
+        },
+    )
+    if response.status_code == 200:
+        count = 0
+        entity_list = []
+        # Get current date and time
+        now = datetime.now()
+        # Get date and time 2 hours later
+        two_hours_later = now + timedelta(hours=2)
+
+        # Format date and time
+        now_str = now.strftime("%Y-%m-%dT%H:%M:%S")
+        two_hours_later_str = two_hours_later.strftime("%Y-%m-%dT%H:%M:%S")
+
+        forecast_list = json.loads(response.content.decode("utf-8"))
+
+        for area in forecast_list["data"]["area_metadata"]:
+            remove_spaced_name = area["name"].replace(" ", "")
+            id = f"{remove_spaced_name}-WeatherForecast-{now_str}_{two_hours_later_str}"
+            entity = Entity("WeatherForecast", id, ctx=ctx)
+
+            for key, value in area.items():
+                if key == "name":
+                    entity.prop("Area", value)
+                    for forecast in forecast_list["data"]["items"][0]["forecasts"]:
+                        if value == forecast["area"]:
+                            entity.prop("forecast", forecast["forecast"])
+                if key == "label_location":
+                    lat = float(area["label_location"]["latitude"])
+                    long = float(area["label_location"]["longitude"])
+                    entity.gprop("location", (lat, long))
+            entity_list.append(entity)
+
+        print(
+            "Total number of forecasts: ",
+            len(forecast_list["data"]["items"][0]["forecasts"]),
+        )
+        print("Total entities created: ", len(entity_list), "\n")
+
+        return entity_list
+    else:
+        return None
+
+# ============================ Weather Observed ======================
 
 # WindDirectionURL = "https://data.gov.sg/datasets?topics=environment&page=1&query=weather+&resultId=d_534cf203023b51f51f879145ccc56ff9"
 # WindSpeedURL = "https://data.gov.sg/datasets?topics=environment&page=1&query=weather+&resultId=d_7677738484067741bf3b56ab5d69c7e9"
 # AirTemperatureURL = "https://data.gov.sg/datasets?topics=environment&page=1&query=weather+&resultId=d_66b77726bbae1b33f218db60ff5861f0"
 # RelativeHumidityURL = "https://data.gov.sg/datasets?topics=environment&page=1&query=weather+&resultId=d_2d3b0c4da128a9a59efca806441e1429#tag/default/GET/air-temperature"
 # RainfallPrecipitationURL = "https://data.gov.sg/datasets?topics=environment&page=1&query=weather+&resultId=d_6580738cdd7db79374ed3152159fbd69#tag/default/GET/air-temperature"
+
 
 dataURL = "https://api-open.data.gov.sg/v2/real-time/api"
 RelativeHumidityURL = dataURL + "/relative-humidity"
@@ -30,7 +100,6 @@ WindDirectionURL = dataURL + "/wind-direction"
 WindSpeedURL = dataURL + "/wind-speed"
 AirTemperatureURL = dataURL + "/air-temperature"
 UVIndexURL = dataURL + "/uv"
-
 
 def fetch_data(url):
     response = requests.get(url)
